notification/requirements.txt
fastapi
uvicorn[standard]
twilio
python-dotenv
